Rework fix_migration_numbers.rake task

+++ lib/tasks/fix_migration_numbers.rake
@@ -11,12 +11,30 @@ namespace :redmine_git_hosting do
       puts "old_name : #{old_name}"
       puts "new_name : #{new_name}"
 
-      query = "DELETE FROM #{ActiveRecord::Base.connection.quote_string('schema_migrations')} WHERE #{ActiveRecord::Base.connection.quote_string('version')} = '#{ActiveRecord::Base.connection.quote_string(old_name)}';"
-      ActiveRecord::Base.connection.execute(query)
+      # Get the old migration name
+      query  = "SELECT * FROM #{ActiveRecord::Base.connection.quote_string('schema_migrations')} WHERE #{ActiveRecord::Base.connection.quote_string('version')} = '#{ActiveRecord::Base.connection.quote_string(old_name)}';"
+      result = ActiveRecord::Base.connection.execute(query)
 
-      query = "INSERT INTO #{ActiveRecord::Base.connection.quote_string('schema_migrations')} (VERSION) VALUES ('#{ActiveRecord::Base.connection.quote_string(new_name)}');"
-      ActiveRecord::Base.connection.execute(query)
+      # If present, rename
+      if !result.to_a.empty?
+        query = "DELETE FROM #{ActiveRecord::Base.connection.quote_string('schema_migrations')} WHERE #{ActiveRecord::Base.connection.quote_string('version')} = '#{ActiveRecord::Base.connection.quote_string(old_name)}';"
+        ActiveRecord::Base.connection.execute(query)
 
+        query = "INSERT INTO #{ActiveRecord::Base.connection.quote_string('schema_migrations')} (VERSION) VALUES ('#{ActiveRecord::Base.connection.quote_string(new_name)}');"
+        ActiveRecord::Base.connection.execute(query)
+      else
+        # Check the new name is present
+        query  = "SELECT * FROM #{ActiveRecord::Base.connection.quote_string('schema_migrations')} WHERE #{ActiveRecord::Base.connection.quote_string('version')} = '#{ActiveRecord::Base.connection.quote_string(new_name)}';"
+        result = ActiveRecord::Base.connection.execute(query)
+
+        if result.to_a.empty?
+          puts "Error : migration is missing #{new_name}"
+        else
+          puts "Already migrated, pass!"
+          puts
+        end
+        next
+      end
       puts
     end
